Rename and update sylvester variants
